pyproject.toml
[tool.poetry]
name = "chatbot-agent-sandbox"
version = "0.1.0"
description = ""
authors = ["Hugo Matalonga <hello@hmatalonga.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
openai = "^0.28.1"
streamlit = ">=1.26"
langchain = "^0.0.306"
python-dotenv = "^1.0.0"
duckduckgo-search = "^3.9.3"
litellm = "^0.11.1"
pyautogen = "^0.1.14"
docker = "^6.1.3"
flask = "^3.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
